LPS-66815 Move classes to new internal structure
