+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-form-web/src/main/java/com/liferay/dynamic/data/lists/form/web/internal/portlet/DDLFormBrowserPortlet.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.dynamic.data.lists.form.web.portlet;
+package com.liferay.dynamic.data.lists.form.web.internal.portlet;
 
-import com.liferay.dynamic.data.lists.form.web.constants.DDLFormPortletKeys;
-import com.liferay.dynamic.data.lists.form.web.display.context.DDLFormBrowserDisplayContext;
+import com.liferay.dynamic.data.lists.form.web.internal.constants.DDLFormPortletKeys;
+import com.liferay.dynamic.data.lists.form.web.internal.display.context.DDLFormBrowserDisplayContext;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetService;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
 import com.liferay.portal.kernel.util.WebKeys;
